requirements.txt
grpcio
grpcio-tools
packaging
pluggy
protobuf
requests
torch
numpy
pillow
opencv-python